pyproject.toml
[tool.poetry]
name = "island"
version = "1.0.0"
description = "Server backend for ClubPenguin"
authors = ["dote", "dan", "ro"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
loguru = "^0.7.0"
jose = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"
pyopenssl = "^23.2.0"
secrets = "^1.0.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.16"}
sqlalchemy-citext = "^1.8.0"
alembic = "^1.11.1"
pynacl = "^1.5.0"
sentry-sdk = {extras = ["loguru"], version = "^1.25.1"}
asyncpg = "^0.27.0"
psycopg2 = "^2.9.6"
passlib = "^1.7.4"
strawberry-graphql = {extras = ["fastapi"], version = "^0.186.1"}
sqlalchemy-utils = "^0.41.1"
redis = "^4.5.5"
fastapi-events = "^0.9.0"
i18nice = {extras = ["yaml"], version = "^0.6.2"}
starlette-context = "^0.3.6"
pydantic = {extras = ["email"], version = "^1.10.9"}

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | env
  | config
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black" 